Fixes, added Alert to handleDeleteMovement(), updated tags in AbstractMovement from @OneToMany to @ManyToMany to fix that some movements would have deleted other movement's tag when inserted in db with those tags.

+++ src/main/java/it/unicam/cs/mpgc/jbudget119250/View/HomeView.java
@@ -244,6 +244,11 @@ public class HomeView implements Initializable {
     private void handleDeleteMovement() {
         Long movement = deleteMovementComboBox.getSelectionModel().getSelectedItem();
 
+        if (movement == null) {
+            AlertView.showAlert("Movement not selected", "Please select a movement from the box", Alert.AlertType.ERROR);
+            return;
+        }
+
         Controller<AbstractMovement> movementController = new DefaultJpaController<>(AbstractMovement.class);
         movementController.delete(movement);
